api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.26"
strawberry-graphql = "^0.219.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
